Change pieces to divs instead of imgs
This makes it so that users can't accidentally highlight the image and drag it around. It also means webpack will pick up the images since they're defined in the CSS, which it wasn't doing before.

Also improves draggability, in case the user moves too fast right after picking up a piece.